modification des IHM

+++ src/classes/CD.java
@@ -7,7 +7,7 @@ public class CD extends Media
 	private int nbChanson;
 	
 	
-	public CD(String titreMedia, String realAutInt, Date dateParution, double prixHT, int dureeLocation,int unNbChanson)
+	public CD(String titreMedia, String realAutInt, Date dateParution, float prixHT, int dureeLocation,int unNbChanson)
 	{
 		super(titreMedia, realAutInt, dateParution, prixHT, dureeLocation);
 		this.nbChanson = unNbChanson;
